+++ app/build.gradle
@@ -60,6 +60,7 @@ android {
 }
 
 dependencies {
+    implementation project(":core")
     implementation project(":event")
     implementation project(':model')
     implementation project(':net:common')
