+++ app/src/main/java/de/danoeh/antennapod/PodcastApp.java
@@ -16,7 +16,7 @@ public class PodcastApp extends Application {
     @Override
     public void onCreate() {
         super.onCreate();
-        Thread.setDefaultUncaughtExceptionHandler(new CrashReportWriter());
+        Thread.setDefaultUncaughtExceptionHandler(new UncaughtExceptionHandler());
         RxJavaErrorHandlerSetup.setupRxJavaErrorHandler();
 
         if (BuildConfig.DEBUG) {
