+++ app/src/main/java/de/danoeh/antennapod/RxJavaErrorHandlerSetup.java
@@ -1,6 +1,8 @@
 package de.danoeh.antennapod;
 
 import android.util.Log;
+
+import de.danoeh.antennapod.core.CrashReportWriter;
 import io.reactivex.rxjava3.exceptions.UndeliverableException;
 import io.reactivex.rxjava3.plugins.RxJavaPlugins;
